requirements.txt
numpy
pandas
tqdm
hydra-core
scikit-learn
scipy
matplotlib
torch